requirements.txt
black
flake8
isort
hydra-core
omegaconf
tinkoff-investments
mkdocs
pip
python-dotenv
-e .
